Add Add to My Reports field.
This field controls whether a report is private to the person who selected
Add to My Reports and saved the report. If that happens, the report is no
longer accessible to any other user, unless that user has the "administer
CiviReport" permission. This also controls whether the report will appear in
the navigation menu or not.