Add Permissions to restrict users from changing the settings on another user's My Reports. Issue #20239.
Anne requested that if a report is private, no other users should be allowed
to edit the settings or change the criteria for the report by saving the
report. They are concerned that someone may inadvertantly "steal" a report by
editing the settings and saving it with Add to My Reports still selected,
which would make the user that saved the settings page the new report owner.