Fix report accordions not working. Issue #20928.
The problem here was that CiviCRM 4.4 handles the javascript for accordions
differently than 4.5+ does. In 4.4, a bit of javascript needs to be added to
each template that accordions are on to initiate the functionality. In 4.5+,
this is all handled without the extra javascript. So, to backport to 4.4, we
have to add that bit in.