CRM-17761
Membership edit shows confirmation and receipt status message even when "Send
Confirmation and Receipt?" is unchecked.

When you edit a membership and leave the "Send Confirmation and Receipt?"
checkbox unchecked, no email sends but Civi still displays the confirmation
message "A confirmation and receipt has been sent to ..."

The fix for now is a simple line adjustment to check for the status of that
checkbox when displaying that message.