(dev/core#285) Fixed second membership repeating reminder
This changes the check for all reminders (not just memberships) to only
check against a sent reminder with the same "reference date". So if we
don't find a sent reminder (ActionLog) with the same reference date as
the one we are trying to send now, then we assume the reminder needs to
be sent again.

This fixes reminders for the cases where case where the original
object's date changes (like membership end dates, for example).

This does change the behavior for some things like events which wouldn't
have reminders sent again if the events date changed. Now if you change
the date of an event and a schedule reminder condition becomes true for
the new date, the reminder will go out even if it was already sent once
for the previous date. That does sound like more of a feature than a bug
to me, but that is a change from previous behavior.